Preprocesseur 3

+++ 0x0D-preprocessor/3-function_like_macro.h
@@ -1,6 +1,6 @@
 #ifndef ABS_FILE
 #define ABS_FILE
 
-#define ABS(a) ((a) < 0 ? (-a) : (a))
+#define ABS(a) ((a) < 0 ? ((unsigned int)-a) : ((unsigned int)a))
 
 #endif
